fix:  suppress traceback in case of exceptions
- Deleted inline `pr.py` and `ticket.py` in `glu/cli`, moving their implementations under `glu/cli/pr` and `glu/cli/ticket` with index exports
- Replaced sprawling `init()` in `main.py` with a single call to `glu.cli.init.init_config`, removing provider, repo and Jira‐template setup helpers
- Cleaned up `main.py` imports, removing unused models/config classes and using annotated index imports for `pr` and `ticket`
- Added `suppress_traceback` decorator in `glu/utils.py` for controlled CLI error handling when `GLU_TEST` is unset
- Reduced cyclomatic complexity and improved modularity of the CLI entrypoint
